feat: main - refactor application structure; collection - add loading functionality; httpclient - implement request handling; ui - create form interface
