+++ internal/logging/logging.go
@@ -1,0 +1,33 @@
+package logging
+
+import (
+	"os"
+
+	"github.com/rs/zerolog"
+	"github.com/rs/zerolog/log"
+)
+
+const (
+	logLevel      = zerolog.WarnLevel
+	logFormatJSON = true
+)
+
+func InitLogger() {
+	zerolog.TimeFieldFormat = zerolog.TimeFormatUnixMs
+	zerolog.SetGlobalLevel(logLevel)
+
+	if logFormatJSON {
+		log.Logger = zerolog.New(os.Stdout).With().Timestamp().Logger()
+	} else {
+		log.Logger = log.Output(zerolog.ConsoleWriter{
+			Out:        os.Stdout,
+			TimeFormat: "2006-01-02 15:04:05.000",
+			NoColor:    false,
+		})
+	}
+}
+
+var (
+	Log     = log.Logger
+	Zerolog = zerolog.Logger{}
+)
